Refactored attribute utils in plugins
